requirements.txt
# local package
-e .

# external requirements
coverage
flake8
python-dotenv>=0.5.1


tensorboard
sklearn
Pillow
tqdm
pandas
numpy

--find-links https://download.pytorch.org/whl/torch_stable.html
torch==1.13.0+cu117
torchvision
